simplify lumi block monitoring
